# docs/api/backend.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShopSteward Backend API — 设计契约（尚未实现）",
    "version": "0.2.0-draft",
    "description": "B0独立业务后端；B2为未来接口。Swagger展示不代表服务已实现。业务语义见../backend-development.md。"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "规划的本地backend地址，当前未启动"
    }
  ],
...
  "paths": {
...
    "/api/v1/dashboard": {
      "get": {
        "operationId": "get_dashboard",
        "tags": [
          "Dashboard"
        ],
        "summary": "查询经营看板，不触发检查",
        "description": "查询经营看板，不触发检查",
        "x-phase": "B0",
        "x-implementation-status": "planned",
        "x-permissions": [
          "viewer"
        ],
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "不透明资源 ID",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                },
                "example": {
                  "state": {
                    "store_id": "store_001",
                    "state_version": 7,
                    "currency": "CNY",
                    "cash_minor": 100000,
                    "reserved_cash_minor": 0,
                    "available_cash_minor": 100000,
                    "receivables_minor": 0,
                    "stocks": [
                      {
                        "sku_id": "sku_001",
                        "on_hand": 20,
                        "in_transit": 0,
                        "remaining_demand": 60,
                        "forecast_version": "fixed-1"
                      }
                    ],
                    "data_as_of": "2026-09-06T00:00:00Z",
                    "simulation_time": "2026-09-06T00:00:00Z"
                  },
                  "freshness": {
                    "status": "FRESH",
                    "data_as_of": "2026-09-06T00:00:00Z",
                    "last_sync_at": "2026-09-06T00:00:00Z"
                  },
                  "active_mission_count": 1,
                  "active_alert_count": 1,
                  "last_check_at": "2026-09-06T00:00:00Z",
                  "next_check_at": "2026-09-06T00:00:30Z"
                }
              }
            }
          },
          "401": {
            "description": "无有效身份",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无对应权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "字段或业务约束无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "超过频率限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "服务暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/missions/{mission_id}/timeline": {
      "get": {
        "operationId": "list_mission_timeline",
        "tags": [
          "Missions"
        ],
        "summary": "查询可追溯业务历史",
        "description": "查询可追溯业务历史",
        "x-phase": "B0",
        "x-implementation-status": "planned",
        "x-permissions": [
          "viewer",
          "agent:read"
        ],
        "security": [
          {
            "UserBearer": []
          },
          {
            "ServiceBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "不透明资源 ID",
              "minLength": 1,
              "maxLength": 128
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineEntryList"
                },
                "example": {
                  "items": [
                    {
                      "id": "timeline_001",
                      "mission_id": "mission_001",
                      "type": "PLAN_CREATED",
                      "summary": "生成40件采购建议",
                      "actor_type": "SYSTEM",
                      "actor_id": null,
                      "references": [
                        {
                          "type": "plan",
                          "id": "plan_001",
                          "version": "1"
                        }
                      ],
                      "created_at": "2026-09-06T00:00:00Z"
                    }
                  ],
                  "next_cursor": null
                }
              }
            }
          },
          "401": {
            "description": "无有效身份",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无对应权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "字段或业务约束无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "超过频率限制",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "服务暂不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "UserBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "用户身份；角色与门店范围由服务端验证。"
      },
      "ServiceBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "服务身份；Agent无审批权限，simulation仅提交授权事件。"
      }
    },
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "retryable": {
            "type": "boolean"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "retryable",
          "details"
        ],
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "Stock": {
        "type": "object",
        "properties": {
          "sku_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "on_hand": {
            "type": "integer",
            "minimum": 0
          },
          "in_transit": {
            "type": "integer",
            "minimum": 0
          },
          "remaining_demand": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "forecast_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "sku_id",
          "on_hand",
          "in_transit",
          "remaining_demand",
          "forecast_version"
        ],
        "additionalProperties": false
      },
      "State": {
        "type": "object",
        "properties": {
          "store_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "state_version": {
            "type": "integer",
            "minimum": 1
          },
          "currency": {
            "type": "string",
            "enum": [
              "CNY"
            ]
          },
          "cash_minor": {
            "type": "integer",
            "minimum": 0
          },
          "reserved_cash_minor": {
            "type": "integer",
            "minimum": 0
          },
          "available_cash_minor": {
            "type": "integer",
            "minimum": 0
          },
          "receivables_minor": {
            "type": "integer",
            "minimum": 0
          },
          "stocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Stock"
            }
          },
          "data_as_of": {
            "anyOf": [
              {
                "type": "string",
                "description": "RFC3339 UTC 时间",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "simulation_time": {
            "anyOf": [
              {
                "type": "string",
                "description": "RFC3339 UTC 时间",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "store_id",
          "state_version",
          "currency",
          "cash_minor",
          "reserved_cash_minor",
          "available_cash_minor",
          "receivables_minor",
          "stocks",
          "data_as_of",
          "simulation_time"
        ],
        "additionalProperties": false
      },
      "Freshness": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "FRESH",
              "STALE",
              "UNKNOWN"
            ]
          },
          "data_as_of": {
            "anyOf": [
              {
                "type": "string",
                "description": "RFC3339 UTC 时间",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_sync_at": {
            "anyOf": [
              {
                "type": "string",
                "description": "RFC3339 UTC 时间",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "status",
          "data_as_of",
          "last_sync_at"
        ],
        "additionalProperties": false
      },
      "Dashboard": {
        "type": "object",
        "properties": {
          "state": {
            "$ref": "#/components/schemas/State"
          },
          "freshness": {
            "$ref": "#/components/schemas/Freshness"
          },
          "active_mission_count": {
            "type": "integer",
            "minimum": 0
          },
          "active_alert_count": {
            "type": "integer",
            "minimum": 0
          },
          "last_check_at": {
            "anyOf": [
              {
                "type": "string",
                "description": "RFC3339 UTC 时间",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_check_at": {
            "anyOf": [
              {
                "type": "string",
                "description": "RFC3339 UTC 时间",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "state",
          "freshness",
          "active_mission_count",
          "active_alert_count",
          "last_check_at",
          "next_check_at"
        ],
        "additionalProperties": false
      },
...
      "Reference": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "mission",
              "plan",
              "action",
              "event",
              "forecast",
              "artifact",
              "scenario",
              "store",
              "alert"
            ]
          },
          "id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "id"
        ],
        "additionalProperties": false
      },
      "TimelineEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "mission_id": {
            "type": "string",
            "description": "不透明资源 ID",
            "minLength": 1,
            "maxLength": 128
          },
          "type": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "actor_type": {
            "type": "string",
            "enum": [
              "USER",
              "SERVICE",
              "SYSTEM"
            ]
          },
          "actor_id": {
            "anyOf": [
              {
                "type": "string",
                "description": "不透明资源 ID",
                "minLength": 1,
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ]
          },
          "references": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reference"
            }
          },
          "created_at": {
            "type": "string",
            "description": "RFC3339 UTC 时间",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "mission_id",
          "type",
          "summary",
          "actor_type",
          "actor_id",
          "references",
          "created_at"
        ],
        "additionalProperties": false
      },
...
      "TimelineEntryList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineEntry"
            }
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "additionalProperties": false
      },
...
    }
  }
}